+++ som_crawlers/__init__.py
@@ -1,4 +1,3 @@
 ## Constructor of som_crawlers module
-import som_crawlers
-import som_crawlers_task
+from models import *
 from wizard import *
