+++ som_crawlers/security/crawlers_security.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+	<data noupdate="0">
+		<record id="res_groups_som_crawlers_manager" model="res.groups">
+			<field name="name">SomCrawlers / Manager</field>
+		</record>
+		<record id="res_groups_som_crawlers_user" model="res.groups">
+			<field name="name">SomCrawlers / Users</field>
+		</record>
+	</data>
+</openerp>
