+++ som_crawlers/tests/__init__.py
@@ -1,3 +1,2 @@
-##Imports
-from som_crawlers_tests import *
-## @brief Constructor for tests file.
+from som_crawlers_config_tests import *
+from wizard_executar_tasca_tests import *
