+++ som_crawlers/views/som_crawlers_result_view.xml
@@ -1,0 +1,40 @@
+ <!-- view de som crawlers Result-->
+<record model="ir.ui.view" id="view_som_crawlers_result_form">
+    <field name="name">som.crawlers.result.form</field>
+    <field name="model">som.crawlers.result</field>
+    <field name="type">form</field>
+    <field name="priority">16</field>
+    <field name="arch" type="xml">
+        <form string="Resultat de la tasca">
+            <field name="task_id" select='1' readonly ='1' />
+            <field name="data_i_hora_execucio" select='1' readonly ='1' />
+            <field name= "resultat" readonly ='1' />
+            <field name="zip_name" readonly ='1' />
+        </form>
+    </field>
+</record>
+
+    <record model="ir.ui.view" id="view_som_crawlers_result_tree">
+    <field name="name">som.crawlers.result.tree</field>
+    <field name="model">som.crawlers.result</field>
+    <field name="type">tree</field>
+    <field name="priority">16</field>
+    <field name="arch" type="xml">
+        <tree string="Resultat de la tasca" editable="bottom">
+            <field name="task_id" readonly ='1'/>
+            <field name="data_i_hora_execucio" select='1' readonly ='1' />
+            <field name= "resultat" readonly ='1' />
+            <field name="zip_name"   />
+        </tree>
+    </field>
+</record>
+
+    <record id="action_som_crawlers_result_tree" model="ir.actions.act_window">
+    <field name="name">Resultat de les tasques automatiques</field>
+    <field name="res_model">som.crawlers.result</field>
+    <field name="view_type">form</field>
+    <field name="view_mode">tree,form</field>
+    <field name="view_id" ref="view_som_crawlers_result_tree" />
+    <field name="context">{'active_test': False}</field>
+</record>
+<menuitem id="menu_crawlers_result" action="action_som_crawlers_result_tree" parent="som_crawlers.menu_crawlers_task"/>
