+++ som_crawlers/views/som_crawlers_task_view.xml
@@ -25,7 +25,6 @@
                     </group>
                 </form>
             </field>
-            
         </record>
         <record model="ir.ui.view" id="view_som_crawlers_task_tree">
             <field name="name">som.crawlers.task.tree</field>
@@ -51,92 +50,6 @@
             <field name="context">{'active_test': False}</field>
         </record>
         <menuitem id="menu_crawlers_task" action="action_som_crawlers_task_tree" parent="som_crawlers.menu_som_crawlers_base"/>
-
-    <!-- view de som crawlers task step-->
-     <record model="ir.ui.view" id="view_som_crawlers_task_step_form">
-            <field name="name">som.crawlers.task.step.form</field>
-            <field name="model">som.crawlers.task.step</field>
-            <field name="type">form</field>
-            <field name="priority">16</field>
-            <field name="arch" type="xml">
-                <form string="Passos de la tasca">
-                    <field name="name" select='1' />
-                    <field name="sequence" select='1' />
-                    <field name= "task_id" select='1' />
-                    <group colspan="6" col="2" string="Execució tasca">
-                        <field name='function' select='1' />
-                        <field name="params" height="200" nolabel="1" select='2'/>
-                    </group>
-                </form>
-            </field>
-        </record>
-         <record model="ir.ui.view" id="view_som_crawlers_task_step_tree">
-            <field name="name">som.crawlers.task.step.tree</field>
-            <field name="model">som.crawlers.task.step</field>
-            <field name="type">tree</field>
-            <field name="priority">16</field>
-            <field name="arch" type="xml">
-                <tree string="Passos de la tasca" editable="bottom">
-                    <field name="task_id" editable='no'/>
-                    <field name="name" editable='no'/>
-                    <field name="sequence" editable='no'/>
-                    <field name="function" editable='no'/>
-                    <field name="params" editable='no'/>
-                </tree>
-            </field>
-        </record>
-         <record id="action_som_crawlers_task_step_tree" model="ir.actions.act_window">
-            <field name="name">Passos Tasques automatiques</field>
-            <field name="res_model">som.crawlers.task.step</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-            <field name="view_id" ref="view_som_crawlers_task_step_tree" />
-            <field name="context">{'active_test': False}</field>
-        </record>
-    <!--<menuitem id="menu_crawlers_task_step" action="action_som_crawlers_task_step_tree" parent="som_crawlers.menu_crawlers_task"/>-->
-
-    <!-- view de som crawlers Result-->
-
-      <record model="ir.ui.view" id="view_som_crawlers_result_form">
-            <field name="name">som.crawlers.result.form</field>
-            <field name="model">som.crawlers.result</field>
-            <field name="type">form</field>
-            <field name="priority">16</field>
-            <field name="arch" type="xml">
-                <form string="Resultat de la tasca">
-                    <field name="task_id" select='1' readonly ='1' />
-                    <field name="data_i_hora_execucio" select='1' readonly ='1' />
-                    <field name= "resultat" readonly ='1' />
-                    <field name="zip_name" readonly ='1' />
-                </form>
-            </field>
-        </record>
-
-         <record model="ir.ui.view" id="view_som_crawlers_result_tree">
-            <field name="name">som.crawlers.result.tree</field>
-            <field name="model">som.crawlers.result</field>
-            <field name="type">tree</field>
-            <field name="priority">16</field>
-            <field name="arch" type="xml">
-                <tree string="Resultat de la tasca" editable="bottom">
-                    <field name="task_id" readonly ='1'/>
-                    <field name="data_i_hora_execucio" select='1' readonly ='1' />
-                    <field name= "resultat" readonly ='1' />
-                    <field name="zip_name"   />
-                </tree>
-            </field>
-        </record>
-
-         <record id="action_som_crawlers_result_tree" model="ir.actions.act_window">
-            <field name="name">Resultat de les tasques automatiques</field>
-            <field name="res_model">som.crawlers.result</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-            <field name="view_id" ref="view_som_crawlers_result_tree" />
-            <field name="context">{'active_test': False}</field>
-        </record>
-        <menuitem id="menu_crawlers_result" action="action_som_crawlers_result_tree" parent="som_crawlers.menu_crawlers_task"/>
-
     </data>
 </openerp>
