+++ som_crawlers/wizard/__init__.py
@@ -1,4 +1,2 @@
 import wizard_executar_tasca
-import wizard_canviarContrasenya
-import wizard_canviarUsuari
-import wizard_canviarDiesDeMarge
+import wizard_canviar_dies_de_marge
