--- som_crawlers/wizard/wizard_canviarUsuari.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<openerp>
-    <data noupdate="0">
-        <record model="ir.ui.view" id="view_wizard_canviarUsuari_form">
-            <field name="name">wizard.canviarUsuari.form</field>
-            <field name="model">wizard.canviarUsuari</field>
-            <field name="type">form</field>
-            <field name="arch" type="xml">
-                <form string="Canviar nom usuari">
-                    <field name="usuari_nou" string="Nou usuari" />
-                    <button special="cancel" string="Cancelar" icon="gtk-cancel"/>
-                    <button name="canviar_usuari" type="object" icon="gtk-ok" string="Canviar"/>
-                </form>
-            </field>
-        </record>
-        <record model="ir.actions.act_window" id="action_wizard_canviarUsuari">
-            <field name="name">Canviar nom usuari</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">wizard.canviarUsuari</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">form</field>
-            <field name="target">new</field>
-            <field name="view_id" ref="view_wizard_canviarUsuari_form"/>
-        </record>
-        <record id="values_canviarUsuari_form" model="ir.values">
-            <field name="object" eval="1"/>
-            <field name="name">Canviar nom usuari</field>
-            <field name="key2">client_action_multi</field>
-            <field name="key">action</field>
-            <field name="model">som.crawlers.config</field>
-            <field name="value" eval="'ir.actions.act_window,'+str(ref('action_wizard_canviarUsuari'))"/>
-        </record>
-    </data>
-</openerp>
