+++ som_crawlers/wizard/wizard_canviar_dies_de_marge.py
@@ -1,16 +1,16 @@
-## Describes the module(wizard) that modifies the old password for a new one
-#imports
+
+# -*- coding: utf-8 -*-
 from osv import osv, fields
 from tools.translate import _
 # @author Ikram Ahdadouche El Idrissi
 # @author Dalila Jbilou Kouhous
 ## Describes the module and contains the function that changes the password
-class WizardCanviarContrasenya(osv.osv_memory):
+class WizardCanviarDiesDeMarge(osv.osv_memory):
 
     ## Module name
-    _name = 'wizard.canviarContrasenya'
+    _name = 'wizard.canviar.dies.de.marge'
 
-    ## Wizard function that changes the password of a configuration which id is activated
+    ## Wizard function that changes the days of margin of a configuration which id is activated
         # @param self The object pointer
         # @param cursor The database pointer
         # @param uid The current user
@@ -18,7 +18,7 @@ class WizardCanviarContrasenya(osv.osv_memory):
         # @param context None certain data to pass
         # @return ir.actions.act_window_close that closes the wizard window
 
-    def canviar_contrasenya(self,cursor, uid, ids, context=None):
+    def canviar_dies(self,cursor, uid, ids, context=None):
 
         if not context:
             return False
@@ -28,18 +28,18 @@ class WizardCanviarContrasenya(osv.osv_memory):
             ids = [ids]
 
         wiz = self.browse(cursor, uid, ids[0], context=context)
-        contrasenya = wiz.contrasenya_nova
+        days_of_margin = wiz.days_of_margin
 
         active_ids = context.get('active_ids')
         ##Load the module
         classconf = self.pool.get('som.crawlers.config')
         for id in active_ids:
-            classconf.canviar_contrasenya(cursor, uid, id, contrasenya, context=context)
+            classconf.canviar_dies_de_marge(cursor, uid, id, int(days_of_margin), context=context)
 
         return {'type': 'ir.actions.act_window_close'}
 
-    ##Column definition : contrasenya nova, the user puts the new password
+    ##Column definition : dies de marge nous, the user puts the margin days
     _columns = {
-        'contrasenya_nova': fields.char("Contrasenya nova", size = 30, required=True, help="Entra la teva nova contrasenya",)
+        'days_of_margin': fields.char("Dies de marge", size = 10, required=True, help="Introdueix un nombre >= 0 de dies de marge",)
     }
-WizardCanviarContrasenya()
+WizardCanviarDiesDeMarge()
